+++ axual-webclient/application_principal.go
@@ -28,7 +28,7 @@ func (c *Client) CreateApplicationPrincipal(applicationPrincipalRequest [1]Appli
 	if err != nil {
 		return "Error sending POST request for application principal", err
 	}
-	time.Sleep(5 * time.Second) // Principal application can take significant time to apply in Kafka cluster
+	time.Sleep(2 * time.Second) // Principal application can take significant time to apply in Kafka cluster
 	return o, nil
 }
 
@@ -43,7 +43,7 @@ func (c *Client) UpdateApplicationPrincipal(id string, applicationUpdatePrincipa
 	if err != nil {
 		return "Error sending PATCH request for application principal", err
 	}
-	time.Sleep(5 * time.Second) // Principal application can take significant time to apply in Kafka cluster
+	time.Sleep(2 * time.Second) // Principal application can take significant time to apply in Kafka cluster
 	return o, nil
 }